LSNBLDR-108; LSBBLDR-107; LSNBLDR-94; fix it to use RSF 0.7.6 correctly
git-svn-id: https://source.sakaiproject.org/svn/lessonbuilder/trunk@113100 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ lessonbuilder/tool/src/java/org/sakaiproject/lessonbuildertool/service/LessonBuilderAccessService.java
@@ -406,7 +406,8 @@ public class LessonBuilderAccessService {
 					// no copyright enforcement, I don't think
 
 					try {
-						long len = resource.getContentLength();
+					    // following cast is redundant is current kernels, but is needed for Sakai 2.6.1
+						long len = (long)resource.getContentLength();
 						String contentType = resource.getContentType();
 						
 						// 	for url resource type, encode a redirect to the body URL
